+++ monopyly/credit/routes.py
@@ -384,7 +384,9 @@ def add_subtransaction_fields():
         subtransaction_count,
     )
     return render_template(
-        "credit/transaction_form/subtransaction_form.html", subform=new_subform
+        "credit/transaction_form/subtransaction_form.html",
+        subform=new_subform,
+        field_list_optional_member=True,
     )